+++ check_manifest.py
@@ -577,6 +577,7 @@ def check_manifest(source_tree='.', create=False, update=False,
     Returns True if the manifest is fine.
     """
     all_ok = True
+    python = os.path.abspath(python)  # in case it was relative
     with cd(source_tree):
         if not is_package():
             raise Failure('This is not a Python project (no setup.py).')